+++ packages/kumo-relational-client/src/kumo_relational_client/adapters/__init__.py
@@ -3,6 +3,5 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from kumo_relational_client.adapters.relational import KumoRelationalAdapter
-from kumo_relational_client.adapters.tabular import KumoTabularAdapter
 
-__all__ = ['KumoRelationalAdapter', 'KumoTabularAdapter']
+__all__ = ['KumoRelationalAdapter']